outil connection_messages totalement fonctionnel
