+++ composer.json
@@ -17,7 +17,7 @@
     "minimum-stability": "dev",
     "prefer-stable": true,
     "require": {
-        "php": ">=8.2",
+        "php": ">=8.1",
         "ext-curl": "*",
         "ext-dom": "*",
         "ext-fileinfo": "*",
