+++ src/main/core/Resources/modules/administration/connection-messages/index.js
@@ -1,6 +1,8 @@
 import {reducer} from '#/main/core/administration/connection-messages/store'
 import {ConnectionMessagesTool} from '#/main/core/administration/connection-messages/containers/tool'
+import {ConnectionMessagesMenu} from '#/main/core/administration/connection-messages/components/menu'
 export default {
   component: ConnectionMessagesTool,
+  menu: ConnectionMessagesMenu,
   store: reducer
 }
