+++ src/main/core/Resources/modules/administration/connection-messages/store/reducer.js
@@ -1,21 +1,22 @@
-import {FORM_SUBMIT_SUCCESS, makeFormReducer} from "#/main/app/content/form/store";
-import {makeListReducer}                      from "#/main/app/content/list/store";
-import {combineReducers, makeReducer}         from "#/main/app/store/reducer";
-import {selectors}                            from "#/main/core/administration/connection-messages/store/selectors";
-import {makeInstanceAction}                   from "#/main/app/store/actions";
-import {TOOL_LOAD}                            from "#/main/core/tool/store";
+import {makeReducer, combineReducers} from '#/main/app/store/reducer'
+import {makeInstanceAction} from '#/main/app/store/actions'
+import {makeFormReducer} from '#/main/app/content/form/store/reducer'
+import {makeListReducer} from '#/main/app/content/list/store'
+import {FORM_SUBMIT_SUCCESS} from '#/main/app/content/form/store/actions'
+
+import {TOOL_LOAD} from '#/main/core/tool/store/actions'
+import {selectors} from '#/main/core/administration/connection-messages/store/selectors'
 
 const reducer = combineReducers({
-
-  messages: combineReducers({
-    list: makeListReducer(selectors.STORE_NAME+'.messages.list', {}, {
-      invalidated: makeReducer(false, {
-        [FORM_SUBMIT_SUCCESS+'/'+selectors.STORE_NAME+'.messages.current']: () => true,
-        [makeInstanceAction(TOOL_LOAD, 'connection_messages')]: () => true
-      })
-    }),
-    current: makeFormReducer(selectors.STORE_NAME+'.messages.current')
-  })
+  list: makeListReducer(selectors.STORE_NAME+'.list', {}, {
+    invalidated: makeReducer(false, {
+      [FORM_SUBMIT_SUCCESS+'/'+selectors.STORE_NAME+'.current']: () => true,
+      [makeInstanceAction(TOOL_LOAD, selectors.STORE_NAME)]: () => true
+    })
+  }),
+  current: makeFormReducer(selectors.STORE_NAME + '.current')
 })
 
-export { reducer }
+export {
+  reducer
+}
