+++ src/main/core/Resources/modules/administration/parameters/store/index.js
@@ -1,10 +1,7 @@
-
-import {actions} from '#/main/core/administration/parameters/store/actions'
 import {reducer} from '#/main/core/administration/parameters/store/reducer'
 import {selectors} from '#/main/core/administration/parameters/store/selectors'
 
 export {
-  actions,
   reducer,
   selectors
 }
