+++ src/main/core/Resources/modules/administration/parameters/store/selectors.js
@@ -1,9 +1,6 @@
 import {createSelector} from 'reselect'
 
-import {selectors as formSelectors} from '#/main/app/content/form/store/selectors'
-
 const STORE_NAME = 'main_settings'
-const FORM_NAME = STORE_NAME+'.parameters'
 
 const store = (state) => state[STORE_NAME]
 
@@ -17,7 +14,6 @@ const lockedParameters = createSelector(
   (store) => store.lockedParameters
 )
 
-const parameters = (state) => formSelectors.data(formSelectors.form(state, FORM_NAME))
 
 const locales = createSelector(
   [parameters],
@@ -26,7 +22,6 @@ const locales = createSelector(
 
 export const selectors = {
   STORE_NAME,
-  FORM_NAME,
 
   store,
   availableLocales,
